watchdog: ensure watchdog shut down during reboot

+++ release/src/router/rc/watchdog.c
@@ -1625,7 +1625,7 @@ void qos_check()
 void watchdog_check()
 {
 	if (!pids("watchdog")){
-		if(nvram_match("upgrade_fw_status", "0")){
+		if(nvram_match("upgrade_fw_status", "0") && !g_reboot){
 			logmessage("watchdog02", "no watchdog, restarting");
 			kill(1, SIGTERM);
 		}
